pkg: shamir: add forgery (cheating) prevention
The rationale for adding forgery protection is quite complicated, and
stems from a core issue in linear secret sharing schemes outlined by
Pieprzyk and Zhang[1]. Effectively the issue boils down to a dishonest
participant in a secret sharing scheme.

If this participant has a valid share, they can modify the Y values so
that the share will no longer correctly recover the secret (and in
principle the other parties cannot be sure that they were cheated, nor
can they discover who cheated them). But if the participant can obtain
the incorrectly recovered secret, they can privately recover the
original secret without the other share holders present.

The technique to solve this problem (called "Cheating Prevention" in the
literature) is *incredibly* complicated, and involves using several
polynomials to produce multiple shares for each party. Then each party
can only reveal a subset of the "sub-shares" they have, and can use the
rest to check if anyone else is cheating. There are many problems with
this scheme, and it's not clear to me that this is the best way of
solving the problem.

Instead, I decided to just go with the "simple" approach. When creating
the original shares, we generate a new ed25519 key and sign all of the
shares with it. The private key is included as part of the "secret
payload" that is being shared, while the public key and signature are
stored as part of the share payload. If anyone modifies the signature,
then verification will fail. If anyone changes the public key used for
signing, then the forgery will be obvious (all real shares have the same
public key). When extending the shares to create more, we can recover
the private key for the signing so we can mint more compatible keys.

The main difference between this scheme and the one proposed in the
literature is that a dishonest participant can still forge shares, but
they can only do it if they already have access to the secret. In my
view, this attack vector is not useful (and it's also not clear to me
how well the technique in the literature protects against this attack).

[1]: https://dx.doi.org/10.1007/3-540-45450-0_9

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>