pkg: crypto: add ChaCha20-Poly1305 encrypt_decrypt helpers
Add the core of our crypto primitives, which are designed to remove the
need to touch any of the "golang.org/x/crypto" code in higher level
code. In particular we abstract away key and nonce generation, and
provide a packet format that is much nicer to deal with.

Rather than using OpenPGP, we use ChaCha20-Poly1305 because it has
authenticated encryption (PGP doesn't really) and because it has support
for more modern cryptography in Go. In addition there is no need to use
the more complicated packet features of OpenPGP so it would just add
more complicated things to the implementation.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>